pyproject.toml
[tool.poetry]
name = "meldingen"
version = "0.1.0"
description = "The main meldingen project which includes the framework bindings."
authors = [
  "Youri Westerman <y.westerman@amsterdam.nl>",
  "David van Buiten <d.van.buiten@amsterdam.nl>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.1"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic-settings = "^2.1.0"
meldingen-core = {git = "https://github.com/Amsterdam/meldingen-core.git", rev = "feat/user"}
sqlmodel = "^0.0.14"
alembic = "^1.13.1"
dependency-injector = {git = "https://github.com/anton-petrov/python-dependency-injector.git"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pydantic = {extras = ["email"], version = "^2.6.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.26"}
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
asgi-lifespan = "^2.1.0"
black = "^24.1.1"
isort = "^5.13.2"
httpx = "^0.26.0"
mkdocstrings = {version = "^0.24.0", extras = ["python"]}
mkdocs-material = "^9.5.6"
mkdocs-literate-nav = "^0.6.1"
mypy = "^1.8.0"
pytest = "^7.0.0"
pytest-alembic = "^0.10.7"
pytest-asyncio = "^0.23.4"
typer = {extras = ["all"], version = "^0.9.0"}
mypy-baseline = "^0.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py312']
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.env
  | \.venv
)/
'''

[tool.isort]
profile = "black"
line_length = 120
known_third_party = ["meldingen_core"]
known_first_party = ["meldingen"]

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

[tool.mypy-baseline]
baseline_path = "mypy-baseline.txt"
